started adding ability to calibrate and save the state of wifi between launches and the progress of a growing tree.

+++ MyCarbonCapture/DJWifiUsageModel.h
@@ -19,5 +19,5 @@
 @property (nonatomic, readonly) double co2gPerByte;
 @property (nonatomic, readonly) double costPerCO2g;
 
-@property (nonatomic, readonly) int networkUsageKB;
+@property (nonatomic, readonly) long networkUsageKB;
 @end
